requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
mysql==0.0.3
mysqlclient==2.2.7
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
